Add MultiCheckbox widget
