+++ src/index.ts
@@ -3,6 +3,7 @@ import Tree from "./Tree";
 import Char from "./Char";
 import Container from "./Container";
 import ContainerWidget from "./ContainerWidget";
+import NewLine from "./NewLine";
 import Selection from "./Selection";
 import Many2one from "./Many2one";
 import Field from "./Field";
@@ -34,6 +35,7 @@ import Indicator from "./Indicator";
 import Dashboard from "./Dashboard";
 import DashboardItem from "./DashboardItem";
 import Tags from "./Tags";
+import MultiCheckbox from "./MultiCheckbox";
 import {
   Graph,
   GraphAxis,
@@ -58,6 +60,7 @@ export {
   Widget,
   Form,
   Tree,
+  NewLine,
   Boolean,
   One2many,
   Integer,
@@ -99,4 +102,5 @@ export {
   graphProcessor,
   graphFieldUtils,
   Tags,
+  MultiCheckbox,
 };
